Improve performance for canvas drawing
Previously I was using two variables for the forward and backward
canvasses, and at every update they'd be compared and switched.
Now I use a table and only compare and switch the a bool variable,
instead of switching the whole texture.
This anecdotally seems to be more efficient for larger displays.